#304 Fix fleet list update in trade route menu
The list was update only if list fleet count changed.
The bug was due to change in fleets destroying a fleet but also creating a fleet with the same name.
As the list doesn't update, the player assign old reference to a destroyed fleet.
The route doesn't work because the fleet is empty before being destroyed so no ship was affected by the trade route
On reload, the destroyed fleet is not save so the trade route saved a invalid reference